backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
psycopg[binary,pool]==3.2.1
cachetools==5.3.3
python-dotenv==1.0.1